refactor: Extract getAdditionalFiles from npm post-update
